+++ components/ui/app-input.tsx
@@ -139,7 +139,7 @@ const StyledTextField = styled(TextField, {
 }));
 
 // --- Component ---
-export const AppInput: React.FC<AppInputProps> = (props) => {
+export const AppInput = React.forwardRef<HTMLElement, AppInputProps>((props, ref) => {
   const [showPassword, setShowPassword] = useState(false);
 
   if (props.type === "checkbox") {
@@ -147,6 +147,7 @@ export const AppInput: React.FC<AppInputProps> = (props) => {
     return (
       <Checkbox
         {...checkboxProps}
+        inputRef={ref}
         sx={{
           color: BORDER_COLOR,
           "&.Mui-checked": {
@@ -178,6 +179,7 @@ export const AppInput: React.FC<AppInputProps> = (props) => {
       label={label}
       isBgWhite={isBgWhite}
       type={isPassword ? (showPassword ? "text" : "password") : type}
+      inputRef={ref}
       InputProps={{
         startAdornment: startIcon ? (
           <InputAdornment position="start">{startIcon}</InputAdornment>
@@ -202,4 +204,6 @@ export const AppInput: React.FC<AppInputProps> = (props) => {
       {...textFieldProps}
     />
   );
-};
+});
+
+AppInput.displayName = "AppInput";